+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
+- `space` parameter added to `tsave` function
 - `space` parameter added to `tprint` function
 - `space` parameter added to `text2art` function
 ## [5.9] - 2023-03-16